+++ CHANGELOG.md
@@ -11,6 +11,13 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v4.0.0-beta.29 (2023-09-27)
+
+#### :rocket: New Feature
+
+* Support for canceling the execution of the directive `v-async-target`
+* Added a new prop `ssrRendering` `iBlock`
+
 ## v4.0.0-beta.28 (2023-09-26)
 
 #### :rocket: New Feature